CP-14062: address review feedback
- Make OpenSSL a hard requirement on Android; fail CMake configure with
  remediation steps instead of falling back to a JS path that can't run
  once batch-derivation headers unconditionally include OpenSSL.
- Wrap each per-index call in the deriveAddressesBatch JS fallback in
  its own try/catch so one bad xpub no longer rejects the whole batch.
- getExtendedPublicKeysForRange now calls ensureAppReady() before the
  loop so the Avalanche app is opened automatically, matching
  getPublicKeysForRange and getSolanaKeysForRange.
- Restore seedless skip of incremental setNonActiveAccounts dispatches
  in migrateRemainingActiveAccounts; batched dispatches were
  reintroducing redundant downstream balance/XP work.
- Combine LedgerService.range.test.ts into LedgerService.test.ts.
