+++ packages/core-mobile/app/services/ledger/LedgerService.test.ts
@@ -10,7 +10,12 @@ import {
   LEDGER_SCAN_FAILED_TITLE,
   LEDGER_SCAN_FAILED_ALREADY_CONNECTED_MESSAGE
 } from './LedgerBluetoothError'
-import { LedgerAppType, LEDGER_ERROR_CODES } from './types'
+import {
+  LedgerAppType,
+  LEDGER_ERROR_CODES,
+  LedgerDerivationPathType,
+  AvalancheKey
+} from './types'
 
 jest.mock('@ledgerhq/react-native-hw-transport-ble', () => ({
   __esModule: true,
@@ -1008,16 +1013,178 @@ describe('LedgerService', () => {
     })
   })
 
+  describe('getAvalancheKeysForRange', () => {
+    const makeMockAvalancheKey = (index: number): AvalancheKey => ({
+      addresses: {
+        evm: `0xEVM${index}`,
+        avm: `X-avax1avm${index}`,
+        pvm: `P-avax1pvm${index}`,
+        coreEth: `C-avax1coreeth${index}`,
+        btc: `bc1btc${index}`
+      },
+      xpubs: { evm: `xpub_evm_${index}`, avalanche: `xpub_avax_${index}` },
+      publicKeys: [
+        {
+          key: `evmpk${index}`,
+          derivationPath: `m/44'/60'/${index}'/0/0`,
+          curve: 'secp256k1' as any
+        },
+        {
+          key: `avaxpk${index}`,
+          derivationPath: `m/44'/9000'/${index}'/0/0`,
+          curve: 'secp256k1' as any
+        }
+      ]
+    })
+
+    beforeEach(() => {
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      jest.spyOn(Logger, 'info').mockImplementation(() => {})
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      jest.spyOn(Logger, 'error').mockImplementation(() => {})
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      jest.spyOn(Logger, 'warn').mockImplementation(() => {})
+      jest.spyOn(LedgerService, 'openApp').mockResolvedValue(undefined)
+      jest.spyOn(LedgerService, 'waitForApp').mockResolvedValue(undefined)
+    })
+
+    afterEach(() => {
+      jest.restoreAllMocks()
+    })
+
+    it('derives keys for each index from 0 to count-1', async () => {
+      const spy = jest
+        .spyOn(LedgerService, 'getAvalancheKeys')
+        .mockImplementation(async (index: number) =>
+          makeMockAvalancheKey(index)
+        )
+
+      const result = await LedgerService.getAvalancheKeysForRange(
+        3,
+        false,
+        LedgerDerivationPathType.BIP44
+      )
+
+      expect(spy).toHaveBeenCalledTimes(3)
+      expect(spy).toHaveBeenCalledWith(0, false, LedgerDerivationPathType.BIP44)
+      expect(spy).toHaveBeenCalledWith(1, false, LedgerDerivationPathType.BIP44)
+      expect(spy).toHaveBeenCalledWith(2, false, LedgerDerivationPathType.BIP44)
+      expect(result).toHaveLength(3)
+      expect(result[0]?.addresses.evm).toBe('0xEVM0')
+      expect(result[2]?.addresses.evm).toBe('0xEVM2')
+    })
+
+    it('skips failed indices >0 and returns null for them', async () => {
+      jest
+        .spyOn(LedgerService, 'getAvalancheKeys')
+        .mockResolvedValueOnce(makeMockAvalancheKey(0))
+        .mockRejectedValueOnce(new Error('Device error'))
+        .mockResolvedValueOnce(makeMockAvalancheKey(2))
+
+      const result = await LedgerService.getAvalancheKeysForRange(
+        3,
+        false,
+        LedgerDerivationPathType.BIP44
+      )
+
+      expect(result).toHaveLength(3)
+      expect(result[0]).not.toBeNull()
+      expect(result[1]).toBeNull()
+      expect(result[2]).not.toBeNull()
+    })
+
+    it('throws if index 0 fails', async () => {
+      jest
+        .spyOn(LedgerService, 'getAvalancheKeys')
+        .mockRejectedValueOnce(new Error('Device error'))
+
+      await expect(
+        LedgerService.getAvalancheKeysForRange(
+          3,
+          false,
+          LedgerDerivationPathType.BIP44
+        )
+      ).rejects.toThrow('Device error')
+    })
+  })
+
   describe('getSolanaKeysForRange', () => {
     beforeEach(() => {
       // eslint-disable-next-line @typescript-eslint/no-empty-function
       jest.spyOn(Logger, 'info').mockImplementation(() => {})
       // eslint-disable-next-line @typescript-eslint/no-empty-function
       jest.spyOn(Logger, 'error').mockImplementation(() => {})
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      jest.spyOn(Logger, 'warn').mockImplementation(() => {})
     })
 
     afterEach(() => {
-      jest.clearAllMocks()
+      jest.restoreAllMocks()
+    })
+
+    it('derives solana keys for each index', async () => {
+      jest.spyOn(LedgerService, 'openApp').mockResolvedValue(undefined)
+      jest.spyOn(LedgerService, 'waitForApp').mockResolvedValue(undefined)
+      const spy = jest
+        .spyOn(LedgerService, 'getSolanaKeys')
+        .mockImplementation(async (index: number) => [
+          {
+            key: `sol${index}`,
+            derivationPath: `44'/501'/${index}'/0/0`,
+            curve: 'ed25519' as any
+          }
+        ])
+
+      const result = await LedgerService.getSolanaKeysForRange(3)
+
+      expect(spy).toHaveBeenCalledTimes(3)
+      expect(result).toHaveLength(3)
+      expect(result[0]?.[0]?.key).toBe('sol0')
+      expect(result[2]?.[0]?.key).toBe('sol2')
+    })
+
+    it('derives solana keys starting from a custom startIndex', async () => {
+      jest.spyOn(LedgerService, 'openApp').mockResolvedValue(undefined)
+      jest.spyOn(LedgerService, 'waitForApp').mockResolvedValue(undefined)
+      const spy = jest
+        .spyOn(LedgerService, 'getSolanaKeys')
+        .mockImplementation(async (index: number) => [
+          {
+            key: `sol${index}`,
+            derivationPath: `44'/501'/${index}'/0/0`,
+            curve: 'ed25519' as any
+          }
+        ])
+
+      const result = await LedgerService.getSolanaKeysForRange(2, 3)
+
+      expect(spy).toHaveBeenCalledTimes(2)
+      expect(spy).toHaveBeenCalledWith(3, undefined, { skipAppCheck: true })
+      expect(spy).toHaveBeenCalledWith(4, undefined, { skipAppCheck: true })
+      expect(result).toHaveLength(2)
+      expect(result[0]?.[0]?.key).toBe('sol3')
+      expect(result[1]?.[0]?.key).toBe('sol4')
+    })
+
+    it('returns null for failed indices but does not throw', async () => {
+      jest.spyOn(LedgerService, 'openApp').mockResolvedValue(undefined)
+      jest.spyOn(LedgerService, 'waitForApp').mockResolvedValue(undefined)
+      jest
+        .spyOn(LedgerService, 'getSolanaKeys')
+        .mockResolvedValueOnce([
+          {
+            key: 'sol0',
+            derivationPath: "44'/501'/0'/0/0",
+            curve: 'ed25519' as any
+          }
+        ])
+        .mockRejectedValueOnce(new Error('Solana app error'))
+
+      const result = await LedgerService.getSolanaKeysForRange(2)
+
+      expect(result).toHaveLength(2)
+      expect(result[0]).not.toBeNull()
+      expect(result[1]).toBeNull()
     })
 
     it('should stop iterating when signal is aborted', async () => {
