+++ packages/core-mobile/app/services/ledger/deriveAddressesOffline.ts
@@ -157,18 +157,29 @@ export async function deriveAddressesBatch(
       'Native deriveAddressesFromXpubs failed, falling back to JS',
       error
     )
-    // Fallback: derive one at a time using the JS implementation
+    // Fallback: derive one at a time using the JS implementation.
+    // A single bad xpub (or thrown derivePublicKey) must not poison the whole
+    // batch — failed indices are omitted from the map so callers can still use
+    // every account that derived successfully.
     const map = new Map<number, DerivedAddresses>()
     for (let i = 0; i < accountIndices.length; i++) {
-      map.set(
-        accountIndices[i]!,
-        deriveAddressesFromXpub(
-          evmXpub,
-          avalancheXpubs[i]!,
-          isTestnet,
-          accountIndices[i]!
+      const accountIndex = accountIndices[i]!
+      try {
+        map.set(
+          accountIndex,
+          deriveAddressesFromXpub(
+            evmXpub,
+            avalancheXpubs[i]!,
+            isTestnet,
+            accountIndex
+          )
         )
-      )
+      } catch (perIndexError) {
+        Logger.error(
+          `JS fallback derivation failed for account index ${accountIndex}`,
+          perIndexError
+        )
+      }
     }
     return map
   }
